Use old and new input system and have a separate example. Put both on player prefab so you can toggle them to see the behavior
