+++ Assets/Walkthrough/Prefabs/Player.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 106288172138341517}
   - component: {fileID: 106288172138341513}
   - component: {fileID: 106288172138341510}
+  - component: {fileID: 1989244369798542965}
   - component: {fileID: 106288172138341511}
   - component: {fileID: 4924704081033154191}
   - component: {fileID: 213984940}
@@ -70,7 +71,26 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d3baa8a342393e149b7ee342827016a4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _componentIndexCache: 0
   _addedNetworkObject: {fileID: 106288172138341511}
+  _networkObjectCache: {fileID: 106288172138341511}
+  RotateSpeed: 150
+  MoveSpeed: 5
+--- !u!114 &1989244369798542965
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 106288172138341508}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7737acf6ac986ba40aeb31c7dcdb6aa6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _componentIndexCache: 4
+  _addedNetworkObject: {fileID: 106288172138341511}
+  _networkObjectCache: {fileID: 106288172138341511}
   RotateSpeed: 150
   MoveSpeed: 5
 --- !u!114 &106288172138341511
@@ -91,6 +111,13 @@ MonoBehaviour:
   _sceneId: 0
   _sceneNetworkObjects:
   - {fileID: 106288172138341511}
+  _networkBehaviours:
+  - {fileID: 106288172138341510}
+  - {fileID: 4924704081033154191}
+  - {fileID: -6967467201534426853}
+  - {fileID: 4327809432728639008}
+  - {fileID: 1989244369798542965}
+  - {fileID: 3886827997641226798}
   SceneTransformProperties:
     Position: {x: 0.6908932, y: 2.23, z: -10.650206}
     Rotation: {x: 0, y: 0, z: 0, w: 1}
@@ -109,7 +136,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a2836e36774ca1c4bbbee976e17b649c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _componentIndexCache: 1
   _addedNetworkObject: {fileID: 106288172138341511}
+  _networkObjectCache: {fileID: 106288172138341511}
   _compress: 1
   _synchronizeParent: 0
   _interpolation: 2
@@ -173,7 +202,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e8cac635f24954048aad3a6ff9110beb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _componentIndexCache: 2
   _addedNetworkObject: {fileID: 106288172138341511}
+  _networkObjectCache: {fileID: 106288172138341511}
   IgnoredParameters: []
   _animator: {fileID: 213984940}
   _smoothFloats: 1
@@ -192,7 +223,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 47ea58410fc3f73498d205517512177c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _componentIndexCache: 3
   _addedNetworkObject: {fileID: 106288172138341511}
+  _networkObjectCache: {fileID: 106288172138341511}
 --- !u!1 &3008959236327998570
 GameObject:
   m_ObjectHideFlags: 0
@@ -319,7 +352,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: abcac2319fffe1e4484dcbbaa6c00e8f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _componentIndexCache: 5
   _addedNetworkObject: {fileID: 106288172138341511}
+  _networkObjectCache: {fileID: 106288172138341511}
   _text: {fileID: 6676287354785153618}
 --- !u!1 &4666051465567237996
 GameObject:
