+++ Assets/Walkthrough/Scripts/MovingWithNewInputSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7737acf6ac986ba40aeb31c7dcdb6aa6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
